CMF-2767 The BlueZ takes more time to identify that the device is OFF
Reason for change: When GetPairedDevice is called, it further makes 2 different APIs. (1) get the list of devices (2) get the device profile for each of the device in that list.
For the 2nd call, since we set infinite timeout to BlueZ, it takes more time confirm if the device is powered off. This leading to TimeOut at IARM where the timeout is set to 5s (for some reason).
So to handle this, we need to discuss with IARM team to identify the reason behind 5 sec timeout and workout numbers.. But till then i changed the timeout of that 2nd call to 2.5 instead of infinite. This will ensure that if the device is not accessible in 2.5s, it will timeout and ensure that the caller does not get any failure.
Test Procedure: As per PACEXi5-1387
Risks: Medium
Signed-off-by: Karunakaran A <Karunakaran_Amirthalingam@cable.comcast.com>

Upstream SHA1: 0c16e101589c5b388ab5bd2cab627dd5913de5f3

Change-Id: I2393b3fdee4ce70de710fb5f13fafbccbe691c6e