DELIA-23555 : BTRCore - Delay SetConfiguration callback to bluez by 1 sec
Reason for change: Acceptable Generic solution for DELIA-23555 at this moment

Information -
SetConfiguration callback is called in the context of the Connect call

Root cause of crash -
Async nature of lower layer bt-ifce stack i.e. bluez session call enters an invalid
state as the reference to the session with the device is completely unreffered when we
return too quickly from the callback which results in the crash at bluez is an open is still
pending at bluez from the device
Delaying the return to bluez results in the session being unreffed with delay and the
incoming open from the device on avdtp to be processed as we have not returrned and there
is a valid ref in bluez

Remove rdk_logger deinit as it is not required.

Test Procedure:
- Scan/Pair/UnPair/Connect/Disconnect Beats Solo3 and Powerbeats3 20 times
- Execute test case presented in RDK-18578 & RDK-19461 & DELIA-25834
    Each test should be exectued atleast 10 times

Risks: Critical

Change-Id: I7e25314079716e6b2fe5b3ed483acdc21f8ed04c
Signed-off-by: Chandresh Pitty <Chandresh_Pitty@cable.comcast.com>